handle timeseries [ci skip]
